fix: spread apis structure
